refactor: connect PositionsList directly to pipes
